Add event Listener to button
